@@ -1,7 +1,7 @@
 {
 	"name": "get-stream",
 	"version": "6.0.1",
-	"description": "Get a stream as a string, buffer, or array",
+	"description": "Get a stream as a string or buffer",
 	"license": "MIT",
 	"repository": "sindresorhus/get-stream",
 	"funding": "https://github.com/sponsors/sindresorhus",
@@ -10,16 +10,17 @@
 		"email": "sindresorhus@gmail.com",
 		"url": "https://sindresorhus.com"
 	},
+	"type": "module",
+	"exports": "./index.js",
 	"engines": {
-		"node": ">=10"
+		"node": ">=16"
 	},
 	"scripts": {
 		"test": "xo && ava && tsd"
 	},
 	"files": [
 		"index.js",
-		"index.d.ts",
-		"buffer-stream.js"
+		"index.d.ts"
 	],
 	"keywords": [
 		"get",
@@ -34,14 +35,14 @@
 		"consume",
 		"readable",
 		"readablestream",
-		"array",
-		"object"
+		"object",
+		"concat"
 	],
 	"devDependencies": {
-		"@types/node": "^14.0.27",
-		"ava": "^2.4.0",
-		"into-stream": "^5.0.0",
-		"tsd": "^0.13.1",
-		"xo": "^0.24.0"
+		"@types/node": "^20.2.4",
+		"ava": "^5.3.0",
+		"into-stream": "^8.0.0",
+		"tsd": "^0.28.1",
+		"xo": "^0.54.2"
 	}
 }
